Allow access to all configuraiton objects
The secretless broker attempts to read all sbconfig configuration
objects. For CI tests, configuration objects are created in API
groups with unique names to avoid colloisions with other tests.
Roles and Role Bindings are used to give access to the relevant
API group for each test. However as stated previously, when
secretless uses a crd for configuration it attempts to read all
potential configuration objects, but RBAC prevents it from successfully
reading the objects that belong to api groups from other tests.

This commit modifies the test config to allow access to all
test api groups. This is not a good solution, but should improve test
reliability.

A better solution would be for secretless to only read the specific
configuration object that it needs, or only enumerate configuration
objects from a specified api group. Neither of these changes are trivial
as the ideal place for introducing the modifcations is in auto generated
code. Therefore the change will involve architectural changes to
secretless which is beyond the current scope.

Related: cyberark/sidecar-injector#4